+++ front/src/components/Gainburua/components/MugikorNabigazioa/MugikorNabigazioa.tsx
@@ -2,9 +2,9 @@ import React, { useState } from 'react';
 
 import { Link } from 'gatsby';
 
-import HamburgerIcon from './assets/hamburguer.svg';
-import Itxi from './assets/itxi.svg';
-import { KontaktuDatuak } from './KontaktuDatuak';
+import HamburgerIcon from '../../assets/hamburguer.svg';
+import Itxi from '../../assets/itxi.svg';
+import { KontaktuDatuak } from '../../KontaktuDatuak';
 import * as styles from './MugikorNabigazioa.module.scss';
 
 export const MugikorNabigazioa: React.FC = () => {
