+++ front/src/components/Gainburua/components/MugikorNabigazioa/index.ts
@@ -1,0 +1,1 @@
+export { MugikorNabigazioa } from './MugikorNabigazioa';
